+++ src/react-components/src/components/02-molecules/SearchBar/SearchBar.stories.tsx
@@ -1,0 +1,14 @@
+import * as React from "react";
+
+import SearchBar from "./SearchBar";
+import { action } from "@storybook/addon-actions";
+import { withKnobs, text } from "@storybook/addon-knobs";
+
+
+export default {
+  title: "SearchBar",
+  component: SearchBar,
+};
+
+export const searchWithParams = () => <SearchBar searchBarId="searchBarId" dropdownId="dropdownId" dropdownOptions={["Keyword", "Other Keyword"]} searchHandler={action("searched")} dropdownBlurHandler={action("selectChanged")}/>;
+export const searchWithNoParams = () => <SearchBar searchBarId="searchBarId" searchHandler={action("searched")} />;