+++ src/react-components/src/index.ts
@@ -3,4 +3,4 @@ export { default as BodyText } from "./components/01-atoms/Text/Text/BodyText";
 export { default as Breadcrumb } from "./components/02-molecules/Menus/Breadcrumbs/Breadcrumbs";
 export { default as UnderlineLink } from "./components/01-atoms/Links/UnderlineLink";
 export { default as IconLink } from "./components/01-atoms/Links/IconLink";
-export { default as ParamSearchBar } from "./components/02-molecules/SearchBar/ParamSearchBar";
+export { default as SearchBar } from "./components/02-molecules/SearchBar/SearchBar";